refactor(session-log): extract sidebar session log into its own feature
intent(session-log): セッションログを sidebar dialog 以外 (terminal preview) からも使うため独立 feature にする
rejected(sidebar): sidebar 配下に置き続ける案 — 関心が異なる 2 feature が SessionLog* を共有する構造になり責務が混ざる
decision(use-session-log-live): rpcClaudeSessionLog + rpcFsWatch + fsChange debounce を composable に集約し dialog と preview が同じ sessionId で同時に開いても二重 watch を作らせない
decision(filer): rpcFsWatch / rpcFsUnwatch を filer barrel から公開し sidebar 側の重複ラッパーを削除
