+++ apps/renderer/src/features/filer/index.ts
@@ -1,6 +1,6 @@
 export { default as FilerPane } from "./FilerPane.vue";
 export { relDirOf } from "./relDirOf";
-export { rpcFsReadFile, rpcFsReadFileAbsolute } from "./rpc";
+export { rpcFsReadFile, rpcFsReadFileAbsolute, rpcFsUnwatch, rpcFsWatch } from "./rpc";
 export type { FsChangePayload } from "./rpc";
 export { getFileIconUrl, getFolderIconUrl } from "./useFileIcon";
 export { useFsWatchSync } from "./useFsWatchSync";
