+++ apps/renderer/src/features/session-log/SessionLogTimeline.vue
