+++ apps/renderer/src/features/session-log/SessionLogTranscript.vue
