+++ apps/renderer/src/features/session-log/sessionLog.test.ts
