+++ apps/renderer/src/features/session-log/sessionLog.ts
